+++ lib/scala/library-manager/src/test/scala/org/enso/librarymanager/LibraryResolverSpec.scala
@@ -17,7 +17,7 @@ class LibraryResolverSpec
     with EitherValue
     with Inside {
   "LibraryResolver" should {
-    val mainRepo = Repository.make("main", "https://example.com/main").get
+    val mainRepo = Repository("main", "https://example.com/main")
     val parentEdition = Editions.Resolved.Edition(
       parent        = None,
       engineVersion = Some(SemVer(0, 0, 0)),
@@ -31,7 +31,7 @@ class LibraryResolverSpec
           )
       )
     )
-    val customRepo = Repository.make("custom", "https://example.com/custom").get
+    val customRepo = Repository("custom", "https://example.com/custom")
     val currentEdition = Editions.Resolved.Edition(
       parent        = Some(parentEdition),
       engineVersion = None,