+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/components/Manifest.scala
@@ -5,7 +5,7 @@ import java.nio.file.Path
 import cats.Show
 import io.circe.{yaml, Decoder}
 import nl.gn0s1s.bump.SemVer
-import org.enso.distribution.OS
+import org.enso.cli.OS
 import org.enso.editions.SemVerJson._
 import org.enso.runtimeversionmanager.components.Manifest.{
   JVMOption,