+++ lib/scala/testkit/src/main/scala/org/enso/testkit/HasTestDirectory.scala
@@ -1,0 +1,9 @@
+package org.enso.testkit
+
+import java.nio.file.Path
+
+trait HasTestDirectory {
+
+  /** Defines a (usually temporary) that is to be used for this test. */
+  def getTestDirectory: Path
+}